allow user to use different access_token_strategy
